+++ ADC Updater/src/com/kaylerrenslow/armaDialogCreator/updater/ADCUpdater.java
@@ -4,11 +4,6 @@ import com.kaylerrenslow.armaDialogCreator.updater.gui.ADCUpdaterWindow;
 import com.kaylerrenslow.armaDialogCreator.updater.tasks.AdcVersionCheckTask;
 import javafx.application.Application;
 import javafx.application.Platform;
-import javafx.beans.value.ChangeListener;
-import javafx.beans.value.ObservableValue;
-import javafx.concurrent.Task;
-import javafx.concurrent.WorkerStateEvent;
-import javafx.event.EventHandler;
 import javafx.stage.Stage;
 
 import java.io.File;
@@ -22,7 +17,7 @@ public class ADCUpdater extends Application {
 	private static final String ADC_JAR = "adc.jar";
 	private static final File ADC_JAR_SAVE_LOCATION = new File("./" + ADC_JAR);
 	private static final File ADC_DOWNLOAD_JAR_SAVE_LOCATION = new File(".");
-	private static final String JSON_RELEASE_INFO = "https://api.github.com/repos/kayler-renslow/arma-dialog-creator/releases/latest";
+	public static final String JSON_RELEASE_INFO_URL = "https://api.github.com/repos/kayler-renslow/arma-dialog-creator/releases/latest";
 
 	public static final ResourceBundle bundle = ResourceBundle.getBundle("com.kaylerrenslow.armaDialogCreator.updater.UpdaterBundle");
 
@@ -32,36 +27,20 @@ public class ADCUpdater extends Application {
 	}
 
 	private ADCUpdaterWindow window;
-	private final ChangeListener<? super String> taskMessagePropertyListener = new ChangeListener<String>() {
-		@Override
-		public void changed(ObservableValue<? extends String> observable, String oldValue, String newValue) {
-			window.getLblStatus().setText(newValue);
-		}
-	};
-	private final ChangeListener<? super Throwable> taskExceptionPropertyListener = new ChangeListener<Throwable>() {
-		@Override
-		public void changed(ObservableValue<? extends Throwable> observable, Throwable oldValue, Throwable newValue) {
-			window.getLblError().setText("Error (" + newValue.getClass().getSimpleName() + "): " + newValue.getMessage());
-			window.addExitButton(ADCUpdater.bundle.getString("Updater.exit"));
-			newValue.printStackTrace();
-		}
-	};
 
 	@Override
 	public void start(Stage primaryStage) throws Exception {
 		this.window = new ADCUpdaterWindow(primaryStage);
 
 		Thread.currentThread().setName("Arma Dialog Creator Updater - JavaFX Thread");
-		primaryStage.show();
+		window.show();
 		runVersionTask();
 	}
 
 	private void runVersionTask() {
-		loadTask(
-				new AdcVersionCheckTask(ADC_JAR_SAVE_LOCATION,
-						ADC_DOWNLOAD_JAR_SAVE_LOCATION,
-						JSON_RELEASE_INFO
-				), "", event -> {
+		window.loadTask(
+				new AdcVersionCheckTask(ADC_JAR_SAVE_LOCATION, ADC_DOWNLOAD_JAR_SAVE_LOCATION),
+				event -> {
 					boolean error = false;
 					if (!ADC_JAR_SAVE_LOCATION.exists()) {
 						window.getLblError().setText(bundle.getString("Updater.Fail.adc_didnt_save"));
@@ -92,14 +71,5 @@ public class ADCUpdater extends Application {
 		);
 	}
 
-	private void loadTask(Task<?> task, String initStatusText, EventHandler<WorkerStateEvent> succeedEvent) {
-		window.getProgressBar().setProgress(-1);
-		window.getLblStatus().setText(initStatusText);
-		task.exceptionProperty().addListener(taskExceptionPropertyListener);
-		task.messageProperty().addListener(taskMessagePropertyListener);
-		task.setOnSucceeded(succeedEvent);
-		window.getProgressBar().progressProperty().bind(task.progressProperty());
-		new Thread(task).start();
-	}
 
 }